Flyttet på endepunkter og gjort ting med uniformt. Endrer til å alltid returnere behandling istedet for fagsak.
